copyrights, user scope, refactorings

+++ php/impl/data/KeyValueStore.php
@@ -1,4 +1,7 @@
 <?php
+/**
+ * Copyright (c) 2010 jacek.pospychala@gmail.com
+ */
 
 class KeyValueStore {
 	
@@ -11,9 +14,12 @@ class KeyValueStore {
 	
 	private $debug = true;
 	
-	public function KeyValueStore($mysqli, $storeName) {
+	private $userid;
+	
+	public function KeyValueStore($mysqli, $userid, $storeName) {
 		$this->table = $storeName;
 		$this->mysqli = $mysqli;
+		$this->userid = $userid;
 	}
 	
 	/**
@@ -25,11 +31,11 @@ class KeyValueStore {
 	 */
 	public function getPath($path, $key) {
 		if (!empty($key)) {
-			$queryStr = "select p.key, p.value from ".$this->table." p where p.path=? and p.key=?";
+			$queryStr = "select p.key, p.value from ".$this->table." p where userid = ".$this->userid." and p.path=? and p.key=?";
 			$stmt = $this->mysqli->prepare($queryStr);
 			$stmt->bind_param("ss", $path, $key);
 		} else {
-			$queryStr = "select p.key, p.value from ".$this->table." p where p.path=?";
+			$queryStr = "select p.key, p.value from ".$this->table." p where userid = ".$this->userid." and p.path=?";
 			$stmt = $this->mysqli->prepare($queryStr);
 			$stmt->bind_param("s", $path);
 		}
@@ -68,8 +74,8 @@ class KeyValueStore {
 			return false;
 		}
 		
-		$delete = $this->mysqli->prepare("delete from ".$this->table." where path=? and ".$this->table.".key=?");
-		$insert = $this->mysqli->prepare("insert into ".$this->table." (path,".$this->table.".key,value) values (?,?,?)");
+		$delete = $this->mysqli->prepare("delete from ".$this->table." where userid = ".$this->userid." and path=? and ".$this->table.".key=?");
+		$insert = $this->mysqli->prepare("insert into ".$this->table." (userid,path,".$this->table.".key,value) values (?,?,?,?)");
 		if (! empty($filter)) {
 			$keyValue  = $prefsMap[$filter];
 			if (empty($keyValue)) {
@@ -78,13 +84,13 @@ class KeyValueStore {
 			}
 			$delete->bind_param("ss", $path, $filter);
 			$delete->execute();
-			$insert->bind_param("sss", $path, $filter, $keyValue);
+			$insert->bind_param("ssss",$this->userid, $path, $filter, $keyValue);
 			$insert->execute();
 		} else {
 			foreach ($prefsMap as $pkey => $pvalue) {
 				$delete->bind_param("ss", $path, $pkey);
 				$delete->execute();
-				$insert->bind_param("sss", $path, $pkey, $pvalue);
+				$insert->bind_param("ssss", $this->userid, $path, $pkey, $pvalue);
 				$insert->execute();
 			}
 		}
@@ -101,14 +107,14 @@ class KeyValueStore {
 	 */
 	public function delete($path, $key) {
 		if (empty($key)) {
-			$queryStr = "delete from ".$this->table." where path = ?";
+			$queryStr = "delete from ".$this->table." where userid = ".$this->userid." and path = ?";
 			$delete = $this->mysqli->prepare($queryStr);
 			$delete->bind_param("s", $path);
 			$delete->execute();
 			
 			
 		} else {
-			$queryStr = "delete from ".$this->table." where path = ? and prefs.key= ?";
+			$queryStr = "delete from ".$this->table." where userid = ".$this->userid." and path = ? and prefs.key= ?";
 			$delete = $this->mysqli->prepare($queryStr);
 			$delete->bind_param("ss", $path, $key);
 			$delete->execute();
